bad request handling

+++ controllers/accountFunctions.js
@@ -44,50 +44,61 @@ async function getAllAccounts(req, res, next) {
 
 async function getAccount(req, res, next) {
     const result = accounts.find(account => account.accountId === req.params.id);
-    try {
-        res.status(200).send(result);
-    }
-    catch (err) {
-        next(err)
-    }
+    if (result) {
+        try {
+            res.status(200).send(result);
+        }
+        catch (err) {
+            next(err)
+        }
+    } else { res.status(404).send(`account: ${req.params.id} not found`) }
 
 }
 
 async function updateCash(req, res, next) {
-    incDecBalance(req.body.accountId, req.body.amount)
-    fs.writeFileSync(accountsDB, JSON.stringify(modified), 'utf-8');
-    try {
-        res.status(200).send(`Success. Account: ${req.body.accountId} balance updated.`);
-    } catch (err) {
-        next(err);
-    }
-}
-
-async function updateCredit(req, res, next) {
-    let parsedAmount = parseFloat(req.body.amount);
-    if (parsedAmount < 0) {
-        res.status(400).send('invalid: credit cannot be negative');
-    } else {
-        let modified = accounts.map((e) => {
-            if (e.accountId == req.body.accountId) {
-                e.credit = parsedAmount;
-                return e;
-            } else {
-                return e;
-            }
-        })
+    const exists = accounts.find(account => account.accountId === req.params.id);
+    if (exists) {
+        incDecBalance(req.body.accountId, req.body.amount)
         fs.writeFileSync(accountsDB, JSON.stringify(modified), 'utf-8');
         try {
-            res.status(200).send(`Success. Account: ${req.body.accountId} credit updated.`);
+            res.status(200).send(`Success. Account: ${req.body.accountId} balance updated.`);
         } catch (err) {
             next(err);
         }
-    }
+    } else { res.status(404).send(`account ${req.body.accountId} not found`) }
 }
+
+async function updateCredit(req, res, next) {
+    const exists = accounts.find(account => account.accountId === req.params.id);
+    if (exists) {
+        let parsedAmount = parseFloat(req.body.amount);
+        if (parsedAmount < 0) {
+            res.status(400).send('invalid: credit cannot be negative');
+        } else {
+            let modified = accounts.map((e) => {
+                if (e.accountId == req.body.accountId) {
+                    e.credit = parsedAmount;
+                    return e;
+                } else {
+                    return e;
+                }
+            })
+            fs.writeFileSync(accountsDB, JSON.stringify(modified), 'utf-8');
+            try {
+                res.status(200).send(`Success. Account: ${req.body.accountId} credit updated.`);
+            } catch (err) {
+                next(err);
+            }
+        }
+    } else { res.status(404).send(`account ${req.body.accountId} not found`) }
+}
+
 
 
 async function transferFunds(req, res, next) {
     senderIndex = accounts.findIndex((account) => account.accountId === req.body.senderId);
+    recieverIndex = accounts.findIndex((account) => account.accountId === req.body.recieverId);
+    if(senderIndex!=-1 && recieverIndex!=-1){
     if (accounts[senderIndex].balance >= req.body.amount) {
         incDecBalance(req.body.senderId, req.body.amount * (-1));
         incDecBalance(req.body.recieverId, req.body.amount);
@@ -98,8 +109,9 @@ async function transferFunds(req, res, next) {
             next(err);
         }
     } else {
-        res.send('insuffecient funds');
+        res.status(400).send('insuffecient funds');
     }
+} else { res.status(404).send(`account not found`) }
 }
 
 module.exports = {
